restructured project layout
