+++ app/project/[id]/components/FileSidebar.jsx
