--- components/TransitionProvider.jsx
@@ -1,22 +1,0 @@
-"use client"
-
-import { AnimatePresence, motion } from "framer-motion"
-import { usePathname } from "next/navigation"
-
-export default function TransitionProvider({ children }) {
-  const pathname = usePathname()
-
-  return (
-    <AnimatePresence mode="wait">
-      <motion.div
-        key={pathname}
-        initial={{ opacity: 0, y: 10 }}
-        animate={{ opacity: 1, y: 0 }}
-        exit={{ opacity: 0, y: -10 }}
-        transition={{ duration: 0.3, ease: "easeInOut" }}
-      >
-        {children}
-      </motion.div>
-    </AnimatePresence>
-  )
-}
